Refactor user and role models: consolidate into a single file and update relationships
