--- app/db/users/user_roles.py
@@ -1,9 +1,0 @@
-from sqlalchemy import Column, Integer, ForeignKey
-from base import Base
-
-
-class UserRole(Base):
-    __tablename__ = "user_roles"
-
-    user_id = Column(Integer, ForeignKey("users.id"), primary_key=True)
-    role_id = Column(Integer, ForeignKey("roles.id"), primary_key=True)
